+++ Package.swift
@@ -24,11 +24,14 @@ let package = Package(
             ],
             publicHeadersPath: "include",
             cSettings: [
+                .define("NS_BLOCK_ASSERTIONS", .when(configuration: .release)),
+                .define("NDEBUG", .when(configuration: .release)),
                 .headerSearchPath("."),
                 .headerSearchPath("Breadcrumbs"),
                 .headerSearchPath("Client"),
                 .headerSearchPath("Configuration"),
                 .headerSearchPath("Delivery"),
+                .headerSearchPath("FeatureFlags"),
                 .headerSearchPath("Helpers"),
                 .headerSearchPath("include/Bugsnag"),
                 .headerSearchPath("KSCrash"),
@@ -52,6 +55,8 @@ let package = Package(
             path: "BugsnagNetworkRequestPlugin/BugsnagNetworkRequestPlugin",
             publicHeadersPath: "include",
             cSettings: [
+                .define("NS_BLOCK_ASSERTIONS", .when(configuration: .release)),
+                .define("NDEBUG", .when(configuration: .release)),
                 .headerSearchPath("."),
                 .headerSearchPath("include/BugsnagNetworkRequestPlugin"),
             ]