+++ features/steps/app_steps.rb
@@ -115,7 +115,7 @@ def relaunch_crashed_app
   when 'ios'
     # Wait for the app to stop running before relaunching
     step 'the app is not running'
-    Maze.driver.launch_app
+    Maze.driver.activate_app Maze.driver.app_id
   when 'macos'
     sleep 4
     launch_app
@@ -131,8 +131,8 @@ end
 def kill_and_relaunch_app
   case Maze::Helper.get_current_platform
   when 'ios'
-    Maze.driver.close_app
-    Maze.driver.launch_app
+    Maze.driver.terminate_app Maze.driver&.app_id
+    Maze.driver.activate_app Maze.driver.app_id
   when 'macos'
     run_macos_app # This will kill the app if it's running
   when 'watchos'
@@ -161,7 +161,12 @@ def run_macos_app
     Process.waitpid $fixture_pid
   end
   dir = 'features/fixtures/macos/output'
-  exe = "#{dir}/macOSTestApp.app/Contents/MacOS/macOSTestApp"
+  if ENV['RUN_XCFRAMEWORK_APP']
+    exe = "#{dir}/macOSTestAppXcFramework.app/Contents/MacOS/macOSTestAppXcFramework"
+  else
+    exe = "#{dir}/macOSTestApp.app/Contents/MacOS/macOSTestApp"
+  end
+
   system("unzip -qd #{dir} #{dir}/macOSTestApp*.zip", exception: true) unless File.exist? exe
   $fixture_pid = Process.spawn($app_env, exe, %i[err out] => '/dev/null')
 end